+++ docs/content/overview/version-history.md
@@ -11,6 +11,20 @@ weight: 30
 Version History
 ===============
 
+### 0.23.0
+
+* Support .NET 4.5: [#295](https://github.com/mysql-net/MySqlConnector/issues/295).
+* Send client connection attributes: [#293](https://github.com/mysql-net/MySqlConnector/issues/293).
+* Dispose `X509Certificate2` objects (.NET 4.6 and later): [#275](https://github.com/mysql-net/MySqlConnector/issues/275).
+
+### 0.22.0
+
+* Add server certificate validation via `MySqlConnectionStringBuilder.CACertificateFile`: [#280](https://github.com/mysql-net/MySqlConnector/pull/280).
+* Support `sha256_password` authentication: [#281](https://github.com/mysql-net/MySqlConnector/issues/281).
+* Ignore `IOException` in `TryPingAsync`: [#289](https://github.com/mysql-net/MySqlConnector/issues/289).
+* Fix "Aborted connection" server errors when `MySqlConnection` isn't disposed: [#290](https://github.com/mysql-net/MySqlConnector/issues/290).
+* Run integration tests on MySQL Server 5.6, MySQL Server 5.7, MariaDB 10.3, Percona Server 5.7.
+
 ### 0.21.0
 
 * Add `MySqlHelper.EscapeString`: [#277](https://github.com/mysql-net/MySqlConnector/issues/277).